[DAGE-115] RRE Vector Search Doctor refactor
