--- rre-tools/src/rre_tools/approximate_search_evaluator/__init__.py
@@ -1,8 +1,0 @@
-from rre_tools.approximate_search_evaluator.config import Config
-from rre_tools.approximate_search_evaluator.constants import SOLR_SUPPORTED_VERSIONS, ELASTICSEARCH_SUPPORTED_VERSIONS
-
-__all__ = [
-    "Config",
-    "SOLR_SUPPORTED_VERSIONS",
-    "ELASTICSEARCH_SUPPORTED_VERSIONS",
-]
