+++ rre-tools/src/rre_tools/vector_search_doctor/approximate_search_evaluator/__init__.py
@@ -1,0 +1,8 @@
+from rre_tools.vector_search_doctor.approximate_search_evaluator.config import Config
+from rre_tools.vector_search_doctor.approximate_search_evaluator.constants import SOLR_SUPPORTED_VERSIONS, ELASTICSEARCH_SUPPORTED_VERSIONS
+
+__all__ = [
+    "Config",
+    "SOLR_SUPPORTED_VERSIONS",
+    "ELASTICSEARCH_SUPPORTED_VERSIONS",
+]
