+++ rre-tools/src/rre_tools/vector_search_doctor/approximate_search_evaluator/constants.py
