+++ rre-tools/src/rre_tools/vector_search_doctor/embedding_model_evaluator/custom_mteb_tasks/__init__.py
@@ -1,0 +1,7 @@
+from rre_tools.vector_search_doctor.embedding_model_evaluator.custom_mteb_tasks.reranking_task import CustomRerankingTask
+from rre_tools.vector_search_doctor.embedding_model_evaluator.custom_mteb_tasks.retrieval_task import CustomRetrievalTask
+
+__all__ = [
+    "CustomRerankingTask",
+    "CustomRetrievalTask"
+]
