+++ rre-tools/src/rre_tools/vector_search_doctor/embedding_model_evaluator/custom_mteb_tasks/reranking_task.py
@@ -6,8 +6,8 @@ from mteb.abstasks.AbsTask import TaskMetadata
 from mteb.abstasks.AbsTaskReranking import AbsTaskReranking
 from mteb.overview import TASKS_REGISTRY
 
-from rre_tools.embedding_model_evaluator.config import Config
-from rre_tools.embedding_model_evaluator.utils import read_corpus_reranking, read_queries, read_candidates
+from rre_tools.vector_search_doctor.embedding_model_evaluator.config import Config
+from rre_tools.vector_search_doctor.embedding_model_evaluator.utils import read_corpus_reranking, read_queries, read_candidates
 
 log = logging.getLogger(__name__)
