+++ rre-tools/src/rre_tools/vector_search_doctor/embedding_model_evaluator/utils.py
