--- rre-tools/tests/embedding_model_evaluator/conftest.py
@@ -1,10 +1,0 @@
-import pytest
-from pathlib import Path
-import warnings
-
-@pytest.fixture
-def resource_folder():
-    return Path(__file__).parent.parent / "resources"
-
-# Since some warnings are raised at import time, I found this solution
-warnings.filterwarnings("ignore", category=DeprecationWarning)
