+++ rre-tools/tests/test_cross_plataform.py
@@ -7,8 +7,8 @@ from rre_tools.shared.data_store import DataStore
 from rre_tools.shared.models import Document
 from rre_tools.shared.writers.writer_config import WriterConfig
 from rre_tools.shared.writers.quepid_writer import QuepidWriter
-from rre_tools.embedding_model_evaluator.embedding_writer import EmbeddingWriter
-from rre_tools.embedding_model_evaluator.constants import TASKS_NAME_MAPPING
+from rre_tools.vector_search_doctor.embedding_model_evaluator.embedding_writer import EmbeddingWriter
+from rre_tools.vector_search_doctor.embedding_model_evaluator.constants import TASKS_NAME_MAPPING
 from rre_tools.dataset_generator.config import Config as DGConfig
