+++ rre-tools/tests/vector_search_doctor/approximate_search_evaluator/conftest.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+import pytest
+
+
+@pytest.fixture
+def resource_folder():
+    return Path(__file__).parent.parent.parent / "resources" / "approximate_search_evaluator"
