+++ rre-tools/tests/vector_search_doctor/embedding_model_evaluator/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+from pathlib import Path
+
+@pytest.fixture
+def resource_folder():
+    return Path(__file__).parent.parent.parent / "resources"
